+++ Capstone_BCG/sqlnet/model/modules/selection_predict.py
@@ -11,8 +11,8 @@ class SelPredictor(nn.Module):
         super(SelPredictor, self).__init__()
         self.use_ca = use_ca
         self.max_tok_num = max_tok_num
-        self.sel_lstm = nn.LSTM(input_size=N_word, hidden_size=N_h/2,
-                num_layers=N_depth, batch_first=True,
+        self.sel_lstm = nn.LSTM(input_size=int(N_word), hidden_size=int(N_h/2),
+                num_layers=int(N_depth), batch_first=True,
                 dropout=0.3, bidirectional=True)
         if use_ca:
             print("Using column attention on selection predicting")
@@ -20,8 +20,8 @@ class SelPredictor(nn.Module):
         else:
             print("Not using column attention on selection predicting")
             self.sel_att = nn.Linear(N_h, 1)
-        self.sel_col_name_enc = nn.LSTM(input_size=N_word, hidden_size=N_h/2,
-                num_layers=N_depth, batch_first=True,
+        self.sel_col_name_enc = nn.LSTM(input_size=int(N_word), hidden_size=int(N_h/2),
+                num_layers=int(N_depth), batch_first=True,
                 dropout=0.3, bidirectional=True)
         self.sel_out_K = nn.Linear(N_h, N_h)
         self.sel_out_col = nn.Linear(N_h, N_h)
@@ -43,8 +43,8 @@ class SelPredictor(nn.Module):
             for idx, num in enumerate(x_len):
                 if num < max_x_len:
                     att_val[idx, :, num:] = -100
-            att = self.softmax(att_val.view((-1, max_x_len))).view(
-                    B, -1, max_x_len)
+            att = self.softmax(att_val.view((-1, int(max_x_len)))).view(
+                    B, -1, int(max_x_len))
             K_sel_expand = (h_enc.unsqueeze(1) * att.unsqueeze(3)).sum(2)
         else:
             h_enc, _ = run_lstm(self.sel_lstm, x_emb_var, x_len)
